[PLG-2080]: Fix the downcast issue...
